Tidied BotClass to have a default behaviour and parameters, added sandbox reset
Have revised WikiBot class to not need a large number of parameters
passed into it. Instead, user can do things like $bot->minor = true to
mark an edit as minor, $bot->conflict->false to ignore edit conflicts,
and need to do $bot->new = true when writing a new page.
---
The Sandbox Reset is a simple script using the class, which will reset
the sandbox to default if not edited within the last six hours.
This set up on wikinewsie.org to run every 30 minutes.